% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crsave.R
\name{crsave}
\alias{crsave}
\title{Wrapper function to save a ggplot with particular defaults}
\usage{
crsave(
  plot_object,
  path = NULL,
  type = NULL,
  device = NULL,
  base_scale = 1,
  width = NULL,
  ratio = 1.5,
  units = "in",
  dpi = 600,
  ...
)
}
\arguments{
\item{plot_object}{A ggplot object}

\item{path}{File name or file path to save the file; defaults to the working
directory and the \code{plot_object} variable name}

\item{type}{Type of file to save the graphic to; defaults to the path
extension or \code{PNG} if no type or extension given}

\item{device}{Which graphics device to use for saving; uses default devices
for main file types}

\item{base_scale}{Scaling factor for the plot output, written to be
compatible with the theme scaling; factor of 1 corresponds to a 6" x 4"
image}

\item{width}{Width of plot image; default is 6"}

\item{ratio}{Aspect ratio of plot image; default is 1.5}

\item{units}{Units for plot width; default is "in"}

\item{dpi}{Resolution for plot image; default is 600}

\item{...}{Other arguments to be passed to the graphics device function or
\code{\link[ggplot2]{ggsave}}}
}
\description{
Wrapper function to save a ggplot with particular defaults
}
