% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crsave.R
\name{device_select}
\alias{device_select}
\title{Helper function to select default graphics device from file type}
\usage{
device_select(type)
}
\arguments{
\item{type}{Type of file to save the graphic to; defaults to the path
extension or \code{PNG} if no type or extension given}
}
\description{
Helper function to select default graphics device from file type
}
