% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legends.R
\name{draw_key_polygon2}
\alias{draw_key_polygon2}
\title{Modification of legend keys for bar plots}
\usage{
draw_key_polygon2(data, params, size)
}
\arguments{
\item{data}{A single row data frame containing the scaled aesthetics to
display in this key}

\item{params}{A list of additional parameters supplied to the geom.}

\item{size}{Width and height of key in mm.}
}
\value{
A grid grob.
}
\description{
This is an extension of \code{\link[ggplot2]{draw_key_polygon}()} from
\code{\link[ggplot2]{ggplot2}} to modify the spacing between legend keys on
bar graphs and other graphs where the \code{fill} aesthetic is used.
}
\examples{
library(ggplot2)

df <- data.frame(
    x = factor(c("IL6R", "IL8R", "IL6R-Ab", "IL8R-Ab", "IL6R-Ab-IL8R"),
          levels = c("IL6R", "IL8R", "IL6R-Ab", "IL8R-Ab", "IL6R-Ab-IL8R")),
    y = c(0.45, 0.78, 0.61, 0.31, 0.72))
p <- ggplot(df, aes(x, y, fill = x)) +
    geom_col(key_glyph = "polygon2")
}
