% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legends.R
\name{expand0}
\alias{expand0}
\alias{expand1}
\title{Convenience functions to remove space between plot and axes}
\usage{
expand0()

expand1()
}
\description{
\code{expand0()} and \code{expand1()} are a convenience functions that call
\code{expansion(mult = c(0,0))} and \code{expansion(mult = c(0,0.1))},
respectively, to remove space between the plot and the axes. See
\code{\link[ggplot2]{expansion}()} for more information on setting the space
between the plotted data and the axes.
}
\examples{
library(ggplot2)

p <- ggplot(faithfuld, aes(waiting, eruptions, fill = density)) +
    geom_raster() +
    scale_fill_distiller(palette = "YlGn") +
    scale_x_continuous(expand = expand0()) +
    scale_y_continuous(expand = expand0())

df <- data.frame(
    x = factor(c("IL6R", "IL8R", "IL6R-Ab", "IL8R-Ab", "IL6R-Ab-IL8R"),
          levels = c("IL6R", "IL8R", "IL6R-Ab", "IL8R-Ab", "IL6R-Ab-IL8R")),
    y = c(0.45, 0.78, 0.61, 0.31, 0.72))
p <- ggplot(df, aes(x, y, fill = x)) +
    geom_col() +
    scale_y_continuous(expand = expand1())
}
