% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_helpers.R
\name{fix_bold}
\alias{fix_bold}
\title{Reapply the bold font to titles and axes that contain formatted math text}
\usage{
fix_bold(font = "Roboto Regular", symbol = TRUE)
}
\arguments{
\item{font}{Font family that was used in \code{\link{theme_cr}()}, given as
a character vector. The font must be included in the theme fonts; use
\code{\link{font_names}()} to see a list of included fonts and
\code{\link{font_preview}()} to preview a selected font.}

\item{symbol}{Should the theme overwrite the system default Symbol font with
the theme default font? If \code{TRUE}, \code{fix_bold()} will overwrite
the Symbol font with the bolded font, and this change will persist until
the end of the R session or until
\code{\link[systemfonts:register_font]{clear_registry()}} is used.}
}
\value{
A list containing the theme properties.
}
\description{
Formatted math text ignores the \code{face} argument of
\code{\link{element_text}()} and is only styled in bold if the
\code{\link{plotmath}} function \code{bold()} is used. To work around that
limitation for the various titles in the plot, this function applies the font
family that is one step "more bold" to the titles.
}
\details{
The required font will be added to the current R session's font registry if
is had not been already (see
\code{\link[systemfonts:register_font]{registry_fonts()}}). The required font
must be included with the theme.
}
\examples{
library(ggplot2)

df <- data.frame(x = 1:5, y = 1:5, z = c("a", "b", "c", "b", "c"))
p <- ggplot(df, aes(x, y, color = z)) +
  geom_point() +
  labs(
  x = bquote(Delta ~ log[10]("Parameter")),
  y = "Y-Axis Label",
  color = "Legend",
  subtitle = "Subtitle Text",
  caption = "Caption Text",
  tag = "A",
  title = bquote(alpha ~ "Regular Text" ~ 1/3)
)

\dontrun{
p + theme_cr() # Uses default scaling and font
p + theme_cr(font = "Roboto Medium") # Bolder font
}
}
