% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font_preview.R
\name{font_preview}
\alias{font_preview}
\alias{font_names}
\alias{font_styles}
\alias{font_table}
\title{Print information and previews of the fonts}
\usage{
font_preview(font = "Roboto Regular")

font_names()

font_styles(font = "Roboto Regular")

font_table()
}
\arguments{
\item{font}{Which font family to preview; defaults to "Roboto Medium"}
}
\description{
\code{font_names()} displays the list of the fonts currently included in
\code{font_register()}. \code{font_table()} generates a formatted
\code{tibble} with information on the properties of all of the included
fonts. \code{font_styles()} displays the included styles for a selected font.
\code{font_preview()} displays a simple preview of the selected font in the
graphics window. The preview requires the \code{\link[ragg]{ragg}} package
and the RStudio graphics backend to be set to \code{AGG} to display correctly.
}
\section{Table Format}{
 For \code{font_table()}:
\describe{\item{\code{family}}{The font family name, used in the
\code{font_*()} functions} \item{\code{style}}{The included styles for each
font family} \item{\code{weight}}{If the style is bold or normal weight}
\item{\code{italic}}{If the style is italicized or not}
\item{\code{features}}{What font features are included in the registered
fonts (like ligatures, tabular numbers, etc.)}}
}

\examples{
\dontrun{
font_preview(font = "Roboto Light")
}
}
