% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legends.R
\name{heatmap_legend}
\alias{heatmap_legend}
\title{Legend theme and appearance tweaking for heatmaps}
\usage{
heatmap_legend(plot_scale = 1, font = "Roboto Regular")
}
\arguments{
\item{plot_scale}{Scaling factor for the plot components as a whole. A value
of 1 corresponds to a font size of 12 pt and a figure size of 6" by 4"}

\item{font}{Font family to be used in the theme, given as a character vector.
The font must be included in the theme fonts; use
\code{\link{font_names}()} to see a list of included fonts and
\code{\link{font_preview}()} to preview a selected font.}
}
\description{
Changes to details of how heatmap legends are drawn to better match the rest
of the custom theme
}
\examples{
library(ggplot2)

p <- ggplot(faithfuld, aes(waiting, eruptions, fill = density)) +
    geom_raster() +
    scale_fill_distiller(palette = "YlGn", guide = heatmap_legend())
}
