% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legends.R
\name{key_scale}
\alias{key_scale}
\alias{draw_key_polygon2}
\title{Modification of legend keys for bar plots}
\usage{
key_scale(scale = 0.8)

draw_key_polygon2(data, params, size, scale = 0.8)
}
\arguments{
\item{scale}{Scaling factor for the height and width of the legend key within
its allotted space. Defaults to 0.8, which gives the key 80 percent of the
allotted height and width.}

\item{data}{A single row data frame containing the scaled aesthetics to
display in this key}

\item{params}{A list of additional parameters supplied to the geom.}

\item{size}{Width and height of key in mm.}
}
\value{
A grid grob.
}
\description{
\code{draw_key_polygon2()} is an extension of
\code{\link[ggplot2]{draw_key_polygon}()} from \code{\link[ggplot2]{ggplot2}}
to modify the spacing between legend keys on bar graphs and other graphs
where the \code{fill} aesthetic is used. It does this by scaling down the legend
keys within their allotted space.
}
\details{
\code{key_scale()} is a wrapper function to allow the specific scaling of the
legend keys to be modified inside the \code{geom} function.

The core part of this approach that differs from the standard \code{ggplot2}
polygon key is that the width and height are changed to \code{unit(scale, "npc")}
instead of \code{unit(1, "npc")}, where \code{scale} has a default value of \code{0.8}. This
makes the width and height of the fill key take up 80 percent (or the \code{scale}
value, if different) of the key space instead of 100 percent of the key
space. This makes the key fill slightly smaller, giving the appearance of
increased space between the keys.
}
\examples{
library(ggplot2)

df <- data.frame(
    x = factor(c("IL6R", "IL8R", "IL6R-Ab", "IL8R-Ab", "IL6R-Ab-IL8R"),
          levels = c("IL6R", "IL8R", "IL6R-Ab", "IL8R-Ab", "IL6R-Ab-IL8R")),
    y = c(0.45, 0.78, 0.61, 0.31, 0.72))
p <- ggplot(df, aes(x, y, fill = x)) +
    geom_col(key_glyph = "polygon2")
}
