% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal_select.R
\name{pal_check}
\alias{pal_check}
\title{Check if the palette argument is formatted correctly}
\usage{
pal_check(palette = "default")
}
\arguments{
\item{palette}{Character vector of length 1 defining the palette to select
colors from; use \code{\link{pal_names}()} to see a list of available
palettes, default palette is named "default"}
}
\description{
Check if the palette argument is formatted correctly
}
