% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal_preview.R
\name{pal_preview}
\alias{pal_preview}
\alias{pal_names}
\alias{pal_colors}
\alias{pal_table}
\title{Print information and previews of the color palettes}
\usage{
pal_preview(
  palette = names(pal),
  ncol = NULL,
  colors = NULL,
  names = NULL,
  alpha = 1,
  label = NULL,
  columns = NULL
)

pal_names()

pal_colors(palette = names(pal))

pal_table()
}
\arguments{
\item{palette}{Character vector of length 1 defining the palette to select
colors from; use \code{\link{pal_names}()} to see a list of available
palettes}

\item{ncol}{Range of colors to select from \code{palette}; if only one number
is provided, it will be the number of colors selected; can only use one
selection technique}

\item{colors}{Names of colors to select from \code{palette}, given as a
character vector; can only use one selection technique}

\item{names}{Names of species to select from \code{palette} (e.g., "IL6" or
"Tocilizumab"), given as a character vector; can only use one selection
technique}

\item{alpha}{new alpha level in [0,1].  If alpha is \code{NA}, existing alpha
values are preserved. Default is \code{1}. Uses \code{\link[scales]{alpha}()} to
set the alpha value}

\item{label}{What text to use for the color preview labels, options are
\code{color}, \code{hex}, or \code{name}, default is \code{color}}

\item{columns}{How many columns to use for the color preview, defaults to
approximately square}
}
\description{
\code{pal_names()} displays the list of the palette names currently included
in \code{pal}. \code{pal_table()} generates a formatted \code{tibble} with
information on the properties of all of the included palettes.
\code{pal_colors()} displays the palette tibble for the selected palette.
\code{pal_preview()} displays a simple preview of the selected colors in the
graphics window, based on the \code{\link[scales]{show_col}()} function from
the \code{\link[scales]{scales}} package.
}
\section{Table Format}{

For \code{pal_table()}: \describe{\item{\code{Name}}{The palette name, used
in the \code{pal_*()} functions.} \item{\code{Key}}{Logical value, if the
palette is one of the main, standard palettes} \item{\code{Named}}{Logical
value, if the palette table contains a \code{name} column with species
names} \item{\code{Type}}{Factor, what category of information should be
depicted each palette} \item{\code{Length}}{Integer, number of colors in
palette} \item{\code{Brightness}}{Factor, is palette the Standard version,
Light, or Dark}}
}

\examples{
pal_preview("antibodies", label = "name", columns = 2)
pal_preview("receptors", names = c("IL6R", "IL6R-Ab"), label = "hex")
}
