% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal_select.R
\name{pal_select}
\alias{pal_select}
\title{Select from the defined color palettes}
\usage{
pal_select(
  palette = names(pal),
  ncol = NULL,
  colors = NULL,
  names = NULL,
  alpha = 1
)
}
\arguments{
\item{palette}{Character vector of length 1 defining the palette to select
colors from; use \code{\link{pal_names}()} to see a list of available
palettes}

\item{ncol}{Range of colors to select from \code{palette}; if only one number
is provided, it will be the number of colors selected; can only use one
selection technique}

\item{colors}{Names of colors to select from \code{palette}, given as a
character vector; can only use one selection technique}

\item{names}{Names of species to select from \code{palette} (e.g., "IL6" or
"Tocilizumab"), given as a character vector; can only use one selection
technique}

\item{alpha}{new alpha level in [0,1].  If alpha is \code{NA}, existing alpha
values are preserved. Default is \code{1}. Uses \code{\link[scales]{alpha}()} to
set the alpha value}
}
\value{
A named character vector with the hex codes of the selected colors.
Color names are given for \code{ncol} and \code{colors} selections, and
species names are given for \code{names} selection.
}
\description{
This selects hex codes from the palettes defined in the internal \code{pal}
dataset. The list of available palettes can be previewed with
\code{\link{pal_names}()}. Subsets of colors can be selected with the
\code{ncol}, \code{colors}, or \code{names} arguments. Only one selection
method can be used in the function call. If no selection method is provided,
the function will return the full color palette with color names.
}
\examples{
pal_select("receptors")
pal_select("receptors", ncol = 3)
pal_select("receptors", ncol = c(2,5))
pal_select("receptors", colors = c("lightred", "red", "orange"))
pal_select("receptors", names = c("IL8", "IL8R", "IL8R-Ab"))
}
