% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crsave.R
\name{path_create}
\alias{path_create}
\title{Helper function to create path for saving the ggplot file}
\usage{
path_create(plotname, path = NULL, type = NULL)
}
\arguments{
\item{plotname}{Plot object variable name converted to a character vector}

\item{path}{File name or file path to save the file; defaults to the working
directory and the \code{plot_object} variable name}

\item{type}{Type of file to save the graphic to; defaults to the path
extension or \code{PNG} if no type or extension given}
}
\description{
Helper function to create path for saving the ggplot file
}
