% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_cr.R
\name{scale_cr}
\alias{scale_cr}
\alias{scale_color_cr}
\alias{scale_fill_cr}
\title{Custom color palette scale functions for ggplot2}
\usage{
scale_color_cr(
  palette = names(pal),
  ncol = NULL,
  colors = NULL,
  species = NULL,
  alpha = 1,
  ...
)

scale_fill_cr(
  palette = names(pal),
  ncol = NULL,
  colors = NULL,
  species = NULL,
  alpha = 1,
  ...
)
}
\arguments{
\item{palette}{Character vector of length 1 defining the palette to select
colors from; use \code{\link{pal_names}()} to see a list of available
palettes}

\item{ncol}{Range of colors to select from \code{palette}; if only one number
is provided, it will be the number of colors selected; can only use one
selection technique}

\item{colors}{Names of colors to select from \code{palette}, given as a
character vector; can only use one selection technique}

\item{species}{Names of species to select from \code{palette} (e.g., "IL6" or
"Tocilizumab"), given as a character vector; can only use one selection
technique}

\item{alpha}{new alpha level in [0,1].  If alpha is \code{NA}, existing alpha
values are preserved. Default is \code{1}. Uses \code{\link[scales]{alpha}()} to
set the alpha value}

\item{...}{Additional parameters for \code{\link[ggplot2]{discrete_scale}}}
}
\description{
Uses the provided color \code{palette} for the \code{color} or \code{fill}
aesthetics of a ggplot2 object. The function will use
\code{\link[ggplot2]{scale_discrete}} when the colors are selected with
\code{ncol} or \code{colors} and will use \code{\link[ggplot2]{scale_manual}}
when the colors are selected with \code{names}.
}
\examples{
library(ggplot2)

df <- data.frame(x = 1:5, y = 1:5, z = c("IL6", "IL8", "IL6R", "IL6", "IL8"))
p <- ggplot(df, aes(x, y, color = z)) + geom_point()

# Using the colors without name-matching
p + scale_color_cr(palette = "receptors")

# Matching the colors to the species names
p + scale_color_cr(palette = "receptors", species = unique(df$z))

}
