% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal_preview.R
\name{show_col}
\alias{show_col}
\title{Quick matrix preview of colors}
\usage{
show_col(colors, labels, columns)
}
\arguments{
\item{colors}{A character vector of colors}

\item{labels}{A character vector of labels corresponding to \code{colors}}

\item{columns}{Number of columns. If not supplied, tried to be as square as
possible.}
}
\description{
Adapted from \code{\link[scales]{show_col}()} from the
\code{\link[scales]{scales}} package. Modified to allow for custom labels.
}
