% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font_register.R
\name{symbol_register}
\alias{symbol_register}
\title{Register selected font family as the symbol font}
\usage{
symbol_register(font = "Roboto Regular")
}
\arguments{
\item{font}{Font family to be registered as the symbol font, defaults
to "Roboto Medium"}
}
\description{
Register selected font family as the symbol font
}
