% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_custom.R
\name{theme_custom}
\alias{theme_custom}
\title{Base customized theme}
\usage{
theme_custom(base_scale = 1, font_scale = 1)
}
\arguments{
\item{base_scale}{Scaling factor for the plot as a whole. A value of 1
corresponds to a font size of 12 pt.}

\item{font_scale}{Scaling factor for the font as compared to the plot. A
value of 1 will scale the font to the plot size. Larger numbers make the
font large compared to the plot size; smaller numbers make the font small
compared to the plot size.}
}
\value{
A list containing the theme properties.
}
\description{
The default customized theme for figure creation. This is also used as the
base theme for additional modifications, such as for bar graphs and heat
maps.
}
\details{
This builds off of the \code{\link[ggplot2]{theme_gray}} function from the
\code{\link[ggplot2]{ggplot2}} package. It overrides components of the
\code{theme_gray} function with the function-specified components, resulting
in a completely custom theme. It uses the packaged fonts by default.
}
