% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_rangeframe.R
\name{theme_rangeframe}
\alias{theme_rangeframe}
\title{Theme modification when using geom_rangeframe}
\usage{
theme_rangeframe(
  plot_scale = 1,
  font_scale = 1,
  font = "Roboto Regular",
  symbol = TRUE,
  cairo = FALSE
)
}
\arguments{
\item{plot_scale}{Scaling factor for the plot components as a whole. A value
of 1 corresponds to a font size of 12 pt and a figure size of 6" by 4"}

\item{font_scale}{Scaling factor for the font as compared to the plot. A
value of 1 will scale the font to the plot size. Larger numbers make the
font large compared to the plot size; smaller numbers make the font small
compared to the plot size.}

\item{font}{Font family to be used in the theme, given as a character vector.
The font must be included in the theme fonts; use
\code{\link{font_names}()} to see a list of included fonts and
\code{\link{font_preview}()} to preview a selected font.}

\item{symbol}{Should the theme overwrite the system default Symbol font with
the theme default font? If \code{TRUE}, \code{theme_cr()} will overwrite
the Symbol font, and this change will persist until the end of the R
session or until \code{\link[systemfonts:register_font]{clear_registry()}}
is used.}

\item{cairo}{Is a cairographics device being used to save the figure? ("pdf"
and "eps" filetypes) Cairo does not use the font registration from the
\code{\link[systemfonts]{systemfonts}} package, so an alternate cairo-safe
font must be provided and the normal font registration will be skipped.}
}
\description{
Modifies select aspects of theme_cr() to make it work better with the
\code{\link[ggthemes]{geom_rangeframe}} from the
\code{\link[ggthemes]{ggthemes}} package.
}
\examples{
library(ggplot2)
library(ggthemes)

df <- data.frame(x = 1:5, y = 1:5, z = c("a", "b", "c", "b", "c"))
p <- ggplot(df, aes(x, y, color = z)) + geom_point() +
    geom_rangeframe(colour = "gray20")

\dontrun{
p + theme_rangeframe()
}
}
